Update Code 2/2 Affichage 1/1 Fir 1/1 Iir 1/2 Integration 3/3 Mesure 2/2 Lecture

+++ define.h
@@ -30,12 +30,12 @@ typedef struct{
   
   ------------------------------------------------------------*/
 typedef struct{
-    int MaxAcr;
-    int MaxAcir;
-    int MinAcr;
-    int MinAcir;
-    int Dcr;
-    int Dcir;
+    float MaxAcr;
+    float MaxAcir;
+    float MinAcr;
+    float MinAcir;
+    float Dcr;
+    float Dcir;
     int periode;
     int passageParZero;
     int poulsPeriod1;
